requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
faiss-cpu
